tests: split into easy and hard LLM questions
This should enable to see easier which questions
are not expected to work with any LLM.
